combine vis functions into one file
